services/auth/requirements.txt
boto3==1.34.122
botocore==1.34.122
bcrypt==4.1.2
PyJWT==2.8.0
pydantic==2.7.1
python-ulid==2.2.0
aws-lambda-powertools==2.31.0
email-validator==2.1.0.post1
